bot/requirements.txt
appdirs==1.4.3       
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cssselect==1.1.0
docutils==0.15.2
EasyProcess==0.3
fake-useragent==0.1.11
geographiclib==1.50
geopy==2.0.0
idna==2.9
jmespath==0.9.5
lxml==4.5.0
names==0.3.0
numpy==1.18.4
pandas==1.0.3
parse==1.15.0
Pillow==7.2.0
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyVirtualDisplay==1.3.2
requests==2.23.0
requests-html==0.10.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tqdm==4.46.0
urllib3==1.25.9
w3lib==1.21.0
webdriver-manager==2.4.0
websockets==8.1